main/requirements.txt
asgiref==3.9.1
attrs==25.3.0
boto3==1.40.4
botocore==1.40.4
Django==5.0.3
django-cors-headers==4.7.0
django-jalali==7.4.0
django-jalali-date==2.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
